requirements.txt
# minigpt4-backend/requirements.txt
fastapi
uvicorn
pillow
torch
transformers
accelerate
torchvision
